moved close contact and laboratory folders into /patient
